+++ Cargo.toml
@@ -39,7 +39,7 @@ lazy_static = "1.5.0"
 paste = "1.0.15"
 pot = "3.0.1"
 rayon = "1.10.0"
-reqwest = "0.12.12"
+reqwest = { version = "0.12.12", features = ["json"] }
 rkyv = { version = "0.8.9", features = ["hashbrown-0_15", "std"] }
 rocksdb = "0.23.0"
 semver = "1.0.23"
@@ -88,6 +88,7 @@ kailua-build = { path = "build/risczero" }
 kailua-client = { path = "bin/client" }
 kailua-common = { path = "crates/common" }
 kailua-contracts = { path = "crates/contracts" }
+kailua-game = { path = "crates/game" }
 kailua-host = { path = "bin/host" }
 
 # Kona git = "https://github.com/op-rs/kona", tag = "kona-client/v1.0.1"
