+++ bin/client/Cargo.toml
@@ -38,6 +38,7 @@ alloy-primitives = { workspace = true, features = ["map-hashbrown"] }
 kailua-build.workspace = true
 kailua-common.workspace = true
 kailua-contracts.workspace = true
+kailua-game.workspace = true
 
 kona-cli.workspace = true
 kona-derive.workspace = true
