+++ bin/client/src/lib.rs
@@ -17,9 +17,7 @@ pub mod bonsai;
 pub mod boundless;
 pub mod oracle;
 pub mod proof;
-pub mod provider;
 pub mod proving;
-pub mod telemetry;
 pub mod witgen;
 pub mod witness;
 pub mod zkvm;
