+++ bin/host/Cargo.toml
@@ -37,6 +37,7 @@ alloy-eips.workspace = true
 kailua-build.workspace = true
 kailua-client.workspace = true
 kailua-common.workspace = true
+kailua-game.workspace = true
 
 # Kona
 kona-cli.workspace = true
