+++ bin/host/src/main.rs
@@ -17,9 +17,9 @@ use alloy_eips::BlockNumberOrTag;
 use alloy_primitives::B256;
 use anyhow::{anyhow, bail, Context};
 use clap::Parser;
-use kailua_client::provider::OpNodeProvider;
 use kailua_client::proving::ProvingError;
 use kailua_common::boot::StitchedBootInfo;
+use kailua_game::provider::optimism::OpNodeProvider;
 use kailua_host::args::KailuaHostArgs;
 use kailua_host::channel::AsyncChannel;
 use kailua_host::config::generate_rollup_config;
