+++ crates/game/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+name = "kailua-game"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+homepage.workspace = true
+repository.workspace = true
+keywords.workspace = true
+categories.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+async-trait.workspace = true
+bytemuck.workspace = true
+c-kzg.workspace = true
+clap.workspace = true
+futures.workspace = true
+hex.workspace = true
+itertools.workspace = true
+reqwest.workspace = true
+rocksdb.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+tokio-retry.workspace = true
+tracing.workspace = true
+
+alloy = { workspace = true, features = ["full", "kzg", "reqwest"] }
+alloy-rpc-types-beacon.workspace = true
+
+opentelemetry.workspace = true
+opentelemetry_sdk.workspace = true
+opentelemetry-otlp.workspace = true
+
+kailua-build.workspace = true
+kailua-common.workspace = true
+kailua-contracts.workspace = true
+
+kona-genesis.workspace = true
+
+[features]
+devnet = []
