+++ crates/game/src/fault.rs
