+++ crates/game/src/provider/beacon.rs
@@ -12,24 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::retry_res_timeout;
+use crate::{await_tel, retry_res_timeout};
 use alloy::consensus::{Blob, BlobTransactionSidecar};
-use alloy::eips::eip4844::{kzg_to_versioned_hash, BLS_MODULUS, FIELD_ELEMENTS_PER_BLOB};
-use alloy::primitives::{B256, U256};
+use alloy::eips::eip4844::kzg_to_versioned_hash;
+use alloy::primitives::B256;
 use alloy_rpc_types_beacon::sidecar::{BeaconBlobBundle, BlobData};
 use anyhow::{bail, Context};
-use kailua_client::await_tel;
 use opentelemetry::global::tracer;
 use opentelemetry::trace::{FutureExt, TraceContextExt, Tracer};
+use reqwest::Client;
 use serde::de::DeserializeOwned;
 use serde_json::Value;
-use std::ops::{Div, Sub};
 use tracing::debug;
 
 #[derive(Clone, Debug)]
 pub struct BlobProvider {
     pub cl_node_endpoint: String,
-    pub client: reqwest::Client,
+    pub client: Client,
     pub genesis_time: u64,
     pub seconds_per_slot: u64,
 }
@@ -40,7 +39,7 @@ impl BlobProvider {
         let context = opentelemetry::Context::current_with_span(tracer.start("BlobProvider::new"));
 
         let cl_node_endpoint = cl_node_endpoint.trim_end_matches('/').to_owned();
-        let client = reqwest::Client::new();
+        let client = Client::new();
 
         let genesis = await_tel!(
             context,
@@ -85,7 +84,7 @@ impl BlobProvider {
     }
 
     pub async fn client_get<T: DeserializeOwned>(
-        client: &reqwest::Client,
+        client: &Client,
         endpoint: &str,
         path: &str,
     ) -> anyhow::Result<T> {
@@ -157,47 +156,4 @@ pub fn blob_sidecar(blob_data: Vec<Blob>) -> anyhow::Result<BlobTransactionSidec
         proofs.push(proof.to_bytes().into_inner().into());
     }
     Ok(BlobTransactionSidecar::new(blobs, commitments, proofs))
-}
-
-pub fn reverse_bits(index: u128, order_po2: u32) -> u128 {
-    index.reverse_bits() >> (u128::BITS - order_po2)
-}
-
-pub const PRIMITIVE_ROOT_OF_UNITY: U256 = U256::from_limbs([7, 0, 0, 0]);
-// primitive_root = 7
-// bls_mod = 52435875175126190479447740508185965837690552500527637822603658699938581184513
-// pow(primitive_root, (bls_mod - 1) // (2 ** 12), bls_mod)
-// 39033254847818212395286706435128746857159659164139250548781411570340225835782
-pub const FE_ORDER_PO2: u32 = 12;
-
-pub fn root_of_unity(index: usize) -> U256 {
-    let primitive_root_exponent = BLS_MODULUS
-        .sub(U256::from(1))
-        .div(U256::from(FIELD_ELEMENTS_PER_BLOB));
-    let root = PRIMITIVE_ROOT_OF_UNITY.pow_mod(primitive_root_exponent, BLS_MODULUS);
-    let root_exponent = reverse_bits(index as u128, FE_ORDER_PO2);
-    root.pow_mod(U256::from(root_exponent), BLS_MODULUS)
-}
-
-pub fn blob_fe_proof(
-    blob: &Blob,
-    index: usize,
-) -> anyhow::Result<(c_kzg::Bytes48, c_kzg::Bytes32)> {
-    let bytes = root_of_unity(index).to_be_bytes();
-    let z = c_kzg::Bytes32::new(bytes);
-    let c_kzg_blob = c_kzg::Blob::from_bytes(blob.as_slice())?;
-    let settings = alloy::consensus::EnvKzgSettings::default();
-    let (proof, value) = settings.get().compute_kzg_proof(&c_kzg_blob, &z)?;
-
-    let commitment = settings.get().blob_to_kzg_commitment(&c_kzg_blob)?;
-
-    let proof_bytes = proof.to_bytes();
-    if settings
-        .get()
-        .verify_kzg_proof(&commitment.to_bytes(), &z, &value, &proof_bytes)?
-    {
-        Ok((proof_bytes, value))
-    } else {
-        bail!("Generated invalid kzg proof.")
-    }
 }
